NES/SNES video updates and misc tweaks/clean up
* Imagine: Make return values for various Pixmap and ProjectionPlane functions more consistent
* EmuFramework: Add EmuSystem::videoAspectRatioScale()
* NES.emu: Add option to set visible video line ranges for showing overscan
* NES.emu: Add option to crop 8 pixels on sides to make tile loading area less distracting on some games
* Snes9x EX+: Add option to control if extended 239/478 line modes are used
* Snes9x EX+: Skip loading patches from current working directory which is never set

+++ C64.emu/src/main/video.cc
@@ -144,8 +144,8 @@ void video_canvas_refresh(struct video_canvas_s *c, unsigned int xs, unsigned in
 	h *= c->videoconfig->scaley;
 	auto pixView = makePixmapView(c);
 
-	w = std::min(w, pixView.w());
-	h = std::min(h, pixView.h());
+	w = std::min((int)w, pixView.w());
+	h = std::min((int)h, pixView.h());
 
 	plugin.video_canvas_render(c, (uint8_t*)pixView.data(), w, h, xs, ys, xi, yi, pixView.pitchBytes());
 }
